refactor: 폴더 구조 리팩토링 진행(#108)
- features/exams/shared 폴더 생성 -> 공통 컴포넌트로 사용되는 파일
- features/exams/queries 폴더 생성 -> React Query 훅 파일
- 배럴 패턴 적용
