+++ src/features/exams/index.ts
@@ -1,25 +1,25 @@
-export { default as EmptyState } from './components/EmptyState.tsx'
+export { default as EmptyState } from './shared/components/EmptyState.tsx'
 export { default as ExamList } from './components/ExamList.tsx'
 export { ExamColumns } from './components/examConfig.tsx'
 export { mockExamList } from '../../mocks/mockExamList.ts'
 export { default as ExamCreateModal } from './components/ExamCreateModal.tsx'
 export { createExamDropdown } from './utils/createExamDropdown.ts'
 export type { ExamQuestion, ExamQuestionResponse } from './types.ts'
-export { ExamQuestionDetailBody } from './exam-managements/questions//detail-modal/ExamQuestionDetailBody.tsx'
-export { ExamQuestionDetailFooter } from './exam-managements/questions/detail-modal/ExamQuestionDetailFooter.tsx'
-export { ExamQuestionDetailSide } from './exam-managements/questions/detail-modal/ExamQuestionDetailSide.tsx'
-export { default as ExamQuestionDetailModal } from './exam-managements/questions/detail-modal/ExamQuestionDetailModal.tsx'
+export { ExamQuestionDetailBody } from './shared/components/detail-modal/ExamQuestionDetailBody.tsx'
+export { ExamQuestionDetailFooter } from './shared/components/detail-modal/ExamQuestionDetailFooter.tsx'
+export { ExamQuestionDetailSide } from './shared/components/detail-modal/ExamQuestionDetailSide.tsx'
+export { default as ExamQuestionDetailModal } from './shared/components/detail-modal/ExamQuestionDetailModal.tsx'
 export { default as ExamDeploymentsModal } from './components/ExamDeploymentsModal.tsx'
 export { default as ExamPage } from '../../pages/ExamManagementPage.tsx'
 export * from './exam-managements/questions/types.ts'
-export { default as ExamDeletePopupModal } from './exam-managements/questions/detail-modal/ExamDeletePopupModal.tsx'
+export { default as ExamDeletePopupModal } from './exam-managements/questions/components/ExamDeletePopupModal.tsx'
 export { default as DistributionList } from './deployments/components/DistributionList.tsx'
 export { default as SubmissionList } from './submissions/components/SubmissionList.tsx'
 export type { Exam, Distribution, Submission } from './types.ts'
 
-export { useExamDeleteMutation } from './hooks/useExamDeleteMutation.ts'
+export { useExamDeleteMutation } from './queries/useExamDeleteMutation.ts'
 
-export { useExamDeploymentsMutation } from './hooks/useExamDeploymentsMutation.ts'
+export { useExamDeploymentsMutation } from './queries/useExamDeploymentsMutation.ts'
 
 export { examDeploymentsSchema } from './schemas/examDeploymentsSchema.ts'
 export { default as DeploymentHistoryModal } from './components/DeploymentHistoryModal.tsx'
