+++ src/features/exams/queries/useExamDeleteMutation.ts
