+++ src/features/exams/shared/components/detail-modal/ExamQuestionDetailBody.tsx
