+++ src/features/exams/shared/components/detail-modal/ExamQuestionDetailModal.tsx
